2 new lessons!
Modifying an element of a list in a function:

Add 3 to the item at index one of the list.
Store the result back into index one.
Return the list.

Using a list of lists in a function:

Create a function called flatten that takes a single list and
concatenates all the sublists that are part of it into a single list.